JWT Authentication: API: Added JWT Authentication with djangosimplejwt Accesstoken contains is_staff parameter to edentify if user is suppossed to access Data Blacklist old Tokens is activated. Added an CustomAccessToken Serializer to add custom claims Added to Endpoints for the Login and the RefreshToken. Frontend: Added an AuthContext with useContext Hook AuthContext contains loginUser and logoutUser Method. AuthContext sets UserContext on ComponentDidMount with LocalStorage Item.
PrivateRoute:
Added Private Route Component to check whether the User is log  ined renders the requested component or redirects the user to the landing Page.

NoUserRoute:
Added a Route Component to render only components when there is no user logined.(For the login Page).

UseAxios Hook:
Added an UseAxios Hook with a Axios Instance.
Added an Axios Interceptor with checks before a Request is commited if the Refreshtoken is not expired. After that is checks if Accesstoken the  is expired. if the RefreshToken is expired, the user gets loged out. The Token is valid 1 day after its been created. If the AccessToken is expired, the refreshtoken endpoint is called to get a new Access and RefreshToken.

TableComponents and DeviceInDetail:
The Data is Fetched via the useAxios Hook.

Added a landing page which still has to be designed